requirements.txt
boto3==1.18.35
botocore==1.21.35
certifi==2021.5.30
charset-normalizer==2.0.4
dnspython==1.16.0
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
jmespath==0.10.0
paho-mqtt==1.5.1
pymongo==3.12.0
pymssql==2.2.2
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
